backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0
email-validator>=2.1.1
motor==3.3.2
pymongo[srv]==4.6.1
faiss-cpu==1.7.4
numpy==1.26.3
sentence-transformers>=2.6.0
python-multipart==0.0.6
httpx==0.26.0
langgraph==0.2.58
langchain-core>=0.2.27,<0.4.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1